+++ tests/testthat/test_parse.R
@@ -1,3 +1,4 @@
+if (isTRUE(Sys.getenv("NOT_CRAN"))) {
 library(testthat)
 if (interactive()) {
     devtools::load_all()
@@ -91,3 +92,4 @@ test_that("no tagged lines, not_from", {
               expect_equal(current, NULL)
 }
 )
+}